move some storage utils to common file
